Updated display configuration to use U8g2 library, adjusted screen offsets, and modified drawing functions for improved compatibility with SSD1306. Enhanced display tests and corrected pixel color handling.

+++ include/display_helpers.h
@@ -3,11 +3,11 @@
 
 #include "config.h"
 
-// Helper functions for abstracted display positioning
+// Helper functions for abstracted display positioning (U8g2, color: 0=black, 1=white)
 void setCursorActual(int16_t x, int16_t y);
-void drawPixelActual(int16_t x, int16_t y, uint16_t color);
-void drawLineActual(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint16_t color);
-void drawRectActual(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t color);
-void fillRectActual(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t color);
+void drawPixelActual(int16_t x, int16_t y, uint8_t color);
+void drawLineActual(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint8_t color);
+void drawRectActual(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t color);
+void fillRectActual(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t color);
 
 #endif // DISPLAY_HELPERS_H
